feat：去除shiro，使用spring自带的filter

+++ ruoyi-common/src/main/java/com/ruoyi/common/core/domain/entity/SysRole.java
@@ -83,7 +83,7 @@ public class SysRole extends BaseEntity
 
     public static boolean isAdmin(Long roleId)
     {
-        return roleId != null && 1L == roleId;
+        return roleId != null && 1L == roleId.longValue();
     }
 
     public String getDataScope()
